server/auth/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.3.0
PyJWT==2.9.0
pyotp==2.9.0
qrcode==8.2
sqlparse==0.5.3
psycopg2-binary>=2.9
